feat: adaptive frame cap (6.6) + float frame_rate support for analysis modes (6.7)
- VideoProcessor.extract_frames: auto-caps at 60 frames when duration * fps exceeds limit
- frame_rate schema changed to float, path helper uses float-safe key
- analyze endpoint: max(0.1, float(frame_rate)) instead of max(1, int)
- TODO: 6.6 and 6.7 marked complete

+++ core_engine/api/schemas/diagnostics.py
@@ -61,7 +61,7 @@ class DiagnosticResult(BaseModel):
 class VideoInferenceRequest(BaseModel):
     request_id: Optional[str] = None
     video_url: Optional[str] = None
-    frame_rate: int = 1
+    frame_rate: float = 1.0
     analysis_depth: str = "standard"
 
 class UploadVideoResponse(BaseModel):
